refactor: update import paths to use absolute imports and integrate
