+++ src/core/components/settings/AskAiStyle.tsx
@@ -1,9 +1,11 @@
+import { useAskAi } from "@/core/hooks/useAskAi.ts";
+import { AskAiResponse } from "@/types/chaibuilder-editor-props.ts";
+import { Button } from "@/ui/shadcn/components/ui/button";
+import { Skeleton } from "@/ui/shadcn/components/ui/skeleton";
+import { Textarea } from "@/ui/shadcn/components/ui/textarea";
 import { Loader } from "lucide-react";
 import { useEffect, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { AskAiResponse } from "../../../types/chaibuilder-editor-props.ts";
-import { Button, Skeleton, Textarea } from "../../../ui";
-import { useAskAi } from "../../hooks/useAskAi.ts";
 import Countdown from "../Countdown.tsx";
 
 export const AskAIStyles = ({ blockId }: { blockId: string | undefined }) => {
