+++ src/core/components/settings/choices/RangeChoices.tsx
@@ -1,10 +1,10 @@
+import { useCurrentClassByProperty } from "@/core/components/settings/choices/BlockStyle";
+import { DropDownChoices } from "@/core/components/settings/choices/DropdownChoices";
+import { StyleContext } from "@/core/components/settings/choices/StyleContext";
+import { useTailwindClassList } from "@/core/constants/CLASSES_LIST";
+import { MinusIcon, PlusIcon } from "@radix-ui/react-icons";
 import { get, nth } from "lodash-es";
 import { useContext, useMemo } from "react";
-import { MinusIcon, PlusIcon } from "@radix-ui/react-icons";
-import { DropDownChoices } from "./DropdownChoices";
-import { useCurrentClassByProperty } from "./BlockStyle";
-import { useTailwindClassList } from "../../../constants/CLASSES_LIST";
-import { StyleContext } from "./StyleContext";
 
 export const RangeChoices = ({ property, onChange }: any) => {
   const { canReset, canChange } = useContext(StyleContext);
